Delete page, Important task and some additions.
Important task feature has been added.
Deleting with slide has been added to Today, Habit and Journal Page.
Adding favourite feature with slide to Today Page has been added.
Delete page has been made fully functional
Filtering tasks in Today page has been added.

Application logo has been added.
Some adjustments have been made for in-app themes.

+++ planova/lib/pages/habit_edit_page.dart
@@ -119,19 +119,39 @@ class _HabitEditPageState extends State<HabitEditPage> {
   }
 
   void _deleteHabit() async {
-    FirebaseFirestore.instance
-        .collection('habits')
-        .doc(widget.habitId)
-        .delete()
-        .then((value) {
-      Navigator.pushAndRemoveUntil(
-        context,
-        MaterialPageRoute(builder: (context) => const Homes()),
-        (Route<dynamic> route) => false,
-      );
-    }).catchError((error) {
-      // Hata işleme kodları
-    });
+    final user = FirebaseAuth.instance.currentUser;
+
+    if (user != null) {
+      DocumentSnapshot habitDoc = await FirebaseFirestore.instance
+          .collection('habits')
+          .doc(widget.habitId)
+          .get();
+      Map<String, dynamic> habitData = habitDoc.data() as Map<String, dynamic>;
+
+      await FirebaseFirestore.instance.collection('deleted_tasks').add({
+        'name': habitData['name'],
+        'description': habitData['description'],
+        'deletedDate': DateTime.now(),
+        'collection': 'habits',
+        'docId': widget.habitId,
+        'userId': user.uid,
+        'data': habitData,
+      });
+
+      FirebaseFirestore.instance
+          .collection('habits')
+          .doc(widget.habitId)
+          .delete()
+          .then((value) {
+        Navigator.pushAndRemoveUntil(
+          context,
+          MaterialPageRoute(builder: (context) => const Homes()),
+          (Route<dynamic> route) => false,
+        );
+      }).catchError((error) {
+        // Hata işleme kodları
+      });
+    }
   }
 
   Future<void> _sendNotificationToFriend(String friendId) async {
